feat(notes): editor a pantalla completa en el Sheet y archivo con tope de 30
