+++ web-app/src/features/notes/hooks/useArchiveNote.ts
@@ -1,0 +1,42 @@
+import { toast } from 'sonner'
+import { useTranslation } from 'react-i18next'
+
+import { defaultNotes } from '../model/notes'
+import { maxArchivedNotes } from '../model/libraryOfArchivedNotes'
+import { ArchivedNote } from '../model/archivedNote'
+import { useNotesQuery } from './useNotesQuery'
+import { useLibraryOfArchivedNotesQuery } from './useLibraryOfArchivedNotesQuery'
+
+/**
+ * Archiva la nota activa del tablero: la mueve al archivo (FIFO, tope
+ * `maxArchivedNotes`) y deja la nota en blanco. `onArchived` sincroniza el
+ * estado local del editor.
+ */
+export function useArchiveNote(onArchived: (text: string) => void) {
+	const { notes, updateNotes } = useNotesQuery()
+	const { archivedNotes, updateArchivedNotes } = useLibraryOfArchivedNotesQuery()
+	const { t } = useTranslation()
+
+	return () => {
+		if (!notes || notes === '' || notes === '<br>' || !archivedNotes) return
+
+		const newArchivedNote: ArchivedNote = {
+			id: crypto.randomUUID(),
+			note: notes,
+			date: new Date(),
+		}
+
+		const newLibrary = {
+			...archivedNotes,
+			archive: [newArchivedNote, ...archivedNotes.archive].slice(0, maxArchivedNotes),
+		}
+
+		updateNotes(defaultNotes, {
+			onSuccess: () => {
+				updateArchivedNotes(newLibrary)
+				onArchived(defaultNotes)
+				toast.success(t('archived_note.archive_successful_toast'))
+			},
+		})
+	}
+}
