+++ web-app/src/features/notes/model/libraryOfArchivedNotes.ts
@@ -7,3 +7,6 @@ export interface LibraryOfArchivedNotes {
 export const defaultLibraryOfArchivedNotes = {
 	archive: [],
 }
+
+/** Tope de notas archivadas: al superarlo se descarta la más antigua (FIFO). */
+export const maxArchivedNotes = 30
