+++ web-app/src/features/notes/ui/Notes.tsx
@@ -1,13 +1,5 @@
 import { Button } from '@/shared/ui/atoms/button'
-import { ScrollArea } from '@/shared/ui/atoms/scroll-area'
-import {
-	Sheet,
-	SheetContent,
-	SheetDescription,
-	SheetHeader,
-	SheetTitle,
-	SheetTrigger,
-} from '@/shared/ui/molecules/sheet'
+import { Sheet, SheetContent, SheetTitle, SheetTrigger } from '@/shared/ui/molecules/sheet'
 import { useTranslation } from 'react-i18next'
 import { useTheme } from '@/shared/hooks/useTheme'
 import { NoteInput } from './NoteInput'
@@ -16,27 +8,16 @@ export default function Notes() {
 	const { t } = useTranslation()
 	const { column, text: textColor } = useTheme()
 	return (
-		<>
-			<Sheet>
-				<SheetTrigger asChild>
-					<Button variant='link' className={`text-base ${textColor}`}>
-						{t('notes.action_title')}
-					</Button>
-				</SheetTrigger>
-				<SheetContent className={column}>
-					<SheetHeader>
-						<SheetTitle className={textColor}>{t('notes.section_title')}</SheetTitle>
-						<SheetDescription aria-describedby='sheet-description'>
-							{t('notes.description')}
-						</SheetDescription>
-					</SheetHeader>
-					<ScrollArea className='h-full'>
-						<main className='p-2 text-base'>
-							<NoteInput />
-						</main>
-					</ScrollArea>
-				</SheetContent>
-			</Sheet>
-		</>
+		<Sheet>
+			<SheetTrigger asChild>
+				<Button variant='link' className={`text-base ${textColor}`}>
+					{t('notes.action_title')}
+				</Button>
+			</SheetTrigger>
+			<SheetContent className={`${column} p-0`} aria-describedby={undefined}>
+				<SheetTitle className='sr-only'>{t('notes.section_title')}</SheetTitle>
+				<NoteInput />
+			</SheetContent>
+		</Sheet>
 	)
 }
